fix: resolve `is_pox_active`` issue
Note: the test uncovered a strange corner case in the way that this is
computed. If there are only two reward slots for the cycle,
`is_pox_active` will be reported as false. This behavior is unchanged
from pox-2, so we won't fix it in 2.4, but will come back to it later.
This change just stacks more so that there are 3 slots filled.